Add launchSettings.json with development and Docker profiles
Added a new `launchSettings.json` file to configure launch
settings for the project. Introduced the `Genocs.Core.Demo.Worker`
profile for local development with the `Development` environment
and enabled `dotnetRunMessages`. Added a `Docker` profile to
support launching the application in a Docker container. This
provides flexibility for running the application in different
environments.